feat: complete task modules
